Refactoring, documentation of new ALTO char filter
