+++ src/test/java/de/digitalcollections/solrocr/formats/alto/AltoCharFilterTest.java
@@ -1,5 +1,7 @@
 package de.digitalcollections.solrocr.formats.alto;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStreamReader;
@@ -9,27 +11,20 @@ import java.nio.file.Paths;
 import org.apache.commons.io.IOUtils;
 import org.junit.Test;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
-public class AltoCharFilterFactoryTest {
+public class AltoCharFilterTest {
   final Path altoPath = Paths.get("src/test/resources/data/alto.xml");
   final Path hyphenPath = Paths.get("src/test/resources/data/bnl_lunion_1865-04-15.xml");
-  final AltoCharFilterFactory fac = new AltoCharFilterFactory();
-  final AltoStaxFilterFactory staxFac = new AltoStaxFilterFactory();
 
   @Test
   public void stripsDescription() throws IOException {
-    Reader reader = fac.create(new InputStreamReader(new FileInputStream(altoPath.toFile())));
+    Reader reader = new AltoCharFilter(new InputStreamReader(new FileInputStream(altoPath.toFile())));
     String s = IOUtils.toString(reader);
     assertThat(s).doesNotContain("ABBYY");
-    Reader staxReader = staxFac.create(new InputStreamReader(new FileInputStream(altoPath.toFile())));
-    String staxS = IOUtils.toString(staxReader);
-    assertThat(staxS).doesNotContain("ABBYY");
   }
 
   @Test
   public void extractsCorrectText() throws IOException {
-    Reader reader = fac.create(new InputStreamReader(new FileInputStream(altoPath.toFile())));
+    Reader reader = new AltoCharFilter(new InputStreamReader(new FileInputStream(altoPath.toFile())));
     String s = IOUtils.toString(reader).replaceAll("[\n\r]+", " ");
     assertThat(s).doesNotContain("<String");
     assertThat(s).doesNotContain("/>");
@@ -38,7 +33,7 @@ public class AltoCharFilterFactoryTest {
 
   @Test
   public void resolvesHyphenation() throws IOException {
-    Reader reader = fac.create(new InputStreamReader(new FileInputStream(hyphenPath.toFile())));
+    Reader reader = new AltoCharFilter(new InputStreamReader(new FileInputStream(hyphenPath.toFile())));
     String s = IOUtils.toString(reader).replaceAll("\\s+", " ");
     assertThat(s).contains("avec lequel elle murmura :");
     assertThat(s).contains("mon vieil ami, faux espoir !");
